# Avoid unnecessary work in `SimpleTopology` constructor when Connectivity eltype is concrete