Didn't benchmark with mixed Connectivity eltypes, but performance for those cases should be
the same (same code for that case).

Benchmarks:

Master:
```julia
master = @benchmark SimpleTopology(faces) setup=(faces=topology(msh)[deleteat!([1:nfaces(msh, 2);], sort!(sample(1:nfaces(msh, 2), round(Int, nfaces(msh, 2)*0.2); replace=false)))]) seconds=10
```
```
BenchmarkTools.Trial: 10000 samples with 1 evaluation per sample.
 Range (min … max):   99.356 μs …   3.460 ms  ┊ GC (min … max):  0.00% … 95.25%
 Time  (median):     112.651 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   155.806 μs ± 311.846 μs  ┊ GC (mean ± σ):  24.92% ± 11.78%

  █                                                             ▁
  ██▄▁▁▁▁▁▁▁▁▁▇▆▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃▆█ █
  99.4 μs       Histogram: log(frequency) by time       2.55 ms <

 Memory estimate: 1.34 MiB, allocs estimate: 11.
```

PR:
```julia
PR = @benchmark SimpleTopology(faces) setup=(faces=topology(msh)[deleteat!([1:nfaces(msh, 2);], sort!(sample(1:nfaces(msh, 2), round(Int, nfaces(msh, 2)*0.2); replace=false)))]) seconds=10
```
```
BenchmarkTools.Trial: 10000 samples with 1 evaluation per sample.
 Range (min … max):   77.014 μs …   3.144 ms  ┊ GC (min … max):  0.00% … 80.44%
 Time  (median):      89.718 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   135.432 μs ± 319.938 μs  ┊ GC (mean ± σ):  30.72% ± 12.39%

  █                                                             ▁
  ██▁▃▁▁▁▁▁▁▁▁▇▆▄▃▁▅▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄▇█ █
  77 μs         Histogram: log(frequency) by time       2.51 ms <

 Memory estimate: 1.33 MiB, allocs estimate: 7.
```

```julia
BenchmarkTools.judge(median(PR), median(master))
```
```
BenchmarkTools.TrialJudgement:
  time:   -20.36% => improvement (5.00% tolerance)
  memory: -0.78% => invariant (1.00% tolerance)
```