@juliohm 2025-04-23
@halleysfifthinc I changed the `<: DataType` check to `isconcretetype`. Please let me know if you see an issue with that.

Waiting tests to merge...